requirements.txt
Flask
pymongo
python-dotenv
spotipy